Example Data Use For Sample
Here, we explore HR data that lets us know about employees' performance levels and the possibility of them staying or leaving. In this example, we use a model for the employees who are currently with the company to see their probability of leaving so we can take action and retain them.