Example Data Use For Sample
For this self-project, a sample data list of speedaters was used. No prior information was given about the data. So we had to assume thing about it and self-identify to the best of our knowledge; for example, what was the speed dating for? What gender was male, and what gender was female? After answering these questions, we can create a quick one-slide presentation that is easy for the audience to read. This slide excludes unnecessary information and focuses on the usable data that, when presented at a quick glance, gives stakeholders an idea of potential strategy. The goal is that with more accurate headers and data, we can then dig deep into the questions about what age was most attractive? Which age group was less attractive? What age group matched the most? and etc