Sample Data Use For Example
This example focuses on a survival analysis of a Machine Factory/Company. In this scenario, our issue is to estimate the estimated lifetime of all the machinery pieces to best re-allocate our maintenance efforts. Potentially catching machine parts that are about to break and service them before they shut down.