Sample Data Use For Example
In this example dataset, we are looking into a chocolate company. In this sample data, the focus was to understand when sales were most important for the company, meaning when sales are being consumed/purchased the most throughout the year. This can help the company understand when sales are expected to go down based on consumption understanding using a boxplot graph.