Sample Data Use For Example
This sample looks at a merchandising store with thousands of customers. The store wants to understand the persona of its customers. They want to know who are my store's 8 customers and to answer questions like when they visit, how often they visit, and how much they spend on a visit.